Ticket #1649: Prepare for prerelease mc-4.7.0-pre3 (code cleanup)
Try to compile with
{{{
make CFLAGS="-Wall -Werror -Wno-long-long -Wwrite-strings"
}}}

+++ edit/editdraw.c
@@ -387,7 +387,7 @@ edit_draw_this_line (WEdit *edit, long b, long row, long start_col,
 						q)) + edit->start_col;
     if ( option_line_state ) {
         cur_line = edit->start_line + row;
-        if ( cur_line <= edit->total_lines ) {
+        if ( cur_line <= (unsigned int) edit->total_lines ) {
             g_snprintf (line_stat, LINE_STATE_WIDTH + 1, "%7i ", cur_line + 1);
         } else {
             memset(line_stat, ' ', LINE_STATE_WIDTH);
